Resolve re-request review
